BE WIP stepper, starter form, tabsheet, remvoe workflows and forms, etc

+++ kinetiqBP-backend/ui-service/src/main/java/org/kinetiqbp/uiservice/repository/FormDefinitionsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormDefinitionsRepository extends JpaRepository<FormDefinition, Long> {
+public interface FormDefinitionsRepository extends JpaRepository<FormDefinition, Integer> {
 }
